Configure `useScrollPosition` as a hook
This commit configures the `handleScrollProgress` function and its
dependencies to be a hook that we can use throughout the codebase.
