+++ client/src/components/Sidebar/sidebar.module.scss
@@ -4,7 +4,6 @@
 @import "../../styles/definitions/colors";
 
 .root {
-  // width: 144px;
   grid-area: sidebar;
   background-color: $primary;
   padding: 32px 36px 32px 60px;
