+++ client/src/modules/Carousel/styles.module.scss
@@ -1,5 +1,4 @@
 .container {
   max-width: 100vw;
-  // width: calc(100% - 144px);
   overflow: hidden;
 }
